requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
pydirectinput>=1.0.4
pygetwindow>=0.0.9
pywin32>=300
pystray>=0.19.0
pillow>=8.0.0
psutil>=5.8.0
pydantic>=1.9.0
onnxruntime>=1.14.0
tensorflow>=2.12.0
scikit-learn>=1.0.0 